--- examples/breakoutBundle/cartridge/metadat.json
@@ -1,13 +1,0 @@
-{
-    "vmlib_ver": "23_8a5",
-    "dependencies": {
-        "pyved_engine": "23.4a4"
-    },
-    "description": "this is a placeholder for the name of your game",
-    "author": "Aleks",
-    "assets": [],
-    "sounds": [],
-    "cartridge": "pyvBreaker",
-    "game_name": "pyvBreaker",
-    "build_date": "Fri Sep 15 12:57:05 2023"
-}
